Move any_adapter
