+++ include/boost/redis/adapter/any_adapter.hpp
